+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+- Support of subdirectories in nextcloud webdav url. Fixes the problem with tagging methods return a 404 
+
 ## 0.1.3 - 2019-01-09
 ### Added
 - provide credentials in client factory